feat(SPV-878): resolved comments in PR
