+++ src/components/NoRecordsText/index.ts
@@ -1,0 +1,1 @@
+export * from './NoRecordsText.tsx';
