+++ src/components/index.ts
@@ -11,9 +11,11 @@ export * from './DateRangeFilter';
 export * from './DestinationsColumns/DestinationsColumns.tsx';
 export * from './Logo';
 export * from './ModeToggle';
+export * from './NoRecordsText';
 export * from './PaymailsColumns/PaymailColumns.tsx';
 export * from './PaymailDeleteDialog';
 export * from './Profile';
+export * from './Searchbar';
 export * from './ui';
 export * from './ViewDialog';
 export * from './XPubColumns/XPubsColumns.tsx';
